% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{bg2fg}
\alias{bg2fg}
\title{Choose a foreground color (dark or light) to contrast with a background color}
\usage{
bg2fg(bg, dark = "#000000", light = "#F8F8F8")
}
\arguments{
\item{bg}{A vector/array of R colors}

\item{dark}{A single R color; should be a dark color}

\item{light}{A single R color; should be a light color}
}
\value{
A vector/array with the same dimensions as \code{bg}, where every element
is either \code{dark} or \code{light}.
}
\description{
Given a background color, \code{bg2fg()} tells you whether to use a \code{dark} or
\code{light} color in the foreground (i.e., for text), so as to maximize contrast
ratio and thus legibility.
}
\details{
Contrast ratio is a function of the luminance of two colors. \code{bg2fg()}
calculates contrast ratio and luminance according to WCAG 2.x definitions:
\itemize{
\item \url{https://www.w3.org/WAI/GL/wiki/Contrast_ratio}
\item \url{https://www.w3.org/WAI/GL/wiki/Relative_luminance}
}

\code{\link[=plot_color]{plot_color()}} and \code{\link[=print_color]{print_color()}} both use \code{bg2fg()} to choose the
foreground text color for their displays.
}
\examples{
bg2fg(c("red", "black", "white", "purple"))

# note that bg2fg() is used behind-the-scenes in the following two calls
plot_color(c("red", "black", "white", "purple"))
print_color(c("red", "black", "white", "purple"))
}
\seealso{
\code{\link[=col2hex]{col2hex()}} for converting colors to hex strings.
}
