% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{col2hex}
\alias{col2hex}
\title{Convert colors to hex color strings}
\usage{
col2hex(col)
}
\arguments{
\item{col}{A vector/array of R colors. Can be character or numeric.}
}
\value{
A vector/array with the same dimensions as \code{col}, where every
element of \code{col} has been converted to a hex string. If the input is a
raster matrix, then the output will be a raster matrix too.
}
\description{
\code{col2hex()} takes a vector/array of R colors (e.g., \code{c("red", "#fac")}) and
standardizes them as hex strings (e.g., \code{c("#FF0000", "#FFAACC")}). The
standardized hex strings will always be uppercase and will have either 6 or 8
hex digits, depending for each color on whether it's opaque.
}
\details{
This function may be useful for programmatically creating code in a different
language, like HTML/CSS. For example, "violetred" is a valid R color but is
not a valid HTML/CSS color. So you could generate valid code like so:

\if{html}{\out{<div class="sourceCode">}}\preformatted{sprintf("<div style='background:\%s;'></div>", col2hex("violetred"))
}\if{html}{\out{</div>}}

Another use case is when drawing \link[grid:rasterGrob]{raster objects} with
the \code{grid} package. If a raster object is specified with R color names, it
can usually be drawn much faster if first converted to hex strings. Compare,
for example, the rendering time for the following two calls to
\code{grid::grid.raster()}. The latter is noticeably faster in the author's
testing.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# big matrix of color names
cols <- rand_color(1e6, hex = FALSE) |> matrix(ncol = 1e3)

# draw the matrix as a raster object
grid::grid.newpage()
grid::grid.raster(cols)

# draw it again, but first convert color names to hex strings
grid::grid.newpage()
grid::grid.raster(col2hex(cols))
}\if{html}{\out{</div>}}
}
\examples{
col2hex(1:8)
col2hex(c("red", "#fac", "1", NA))
}
\seealso{
\code{\link[=bg2fg]{bg2fg()}} for choosing a foreground color to contrast with a
background color.
}
