% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-color.R
\name{rand_color}
\alias{rand_color}
\title{Random sampling of colors}
\usage{
rand_color(n = 1, hex = 0.5, alpha = FALSE)
}
\arguments{
\item{n}{The number of colors to sample}

\item{hex}{Either \code{TRUE} (meaning sample only hex strings), \code{FALSE} (meaning
sample only color names), or a real number from 0 to 1 indicating the
proportion of colors that should be hex strings (rather than color names)}

\item{alpha}{\code{TRUE}/\code{FALSE} for whether hex strings should have a random
alpha value}
}
\value{
A character vector of length \code{n}
}
\description{
Random sampling of colors
}
