% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-color.R
\name{plot_color}
\alias{plot_color}
\title{Plot colored cells}
\usage{
plot_color(
  col,
  label = TRUE,
  index = TRUE,
  border = graphics::par("bg"),
  cex_label = 1
)
}
\arguments{
\item{col}{A vector, matrix, or list of vectors/matrices, all containing R
colors. Can be character or numeric.}

\item{label}{\code{TRUE}/\code{FALSE} for whether print the color names.}

\item{index}{\code{TRUE}/\code{FALSE} for whether to print row and column indices (as
applicable).}

\item{border}{Border color for each cell. Use \code{NA} for transparent borders,
or \code{NULL} to make borders match the cell background colors.}

\item{cex_label}{Size of printed color names, as multiplier of default size.}
}
\value{
Invisibly returns a \link[gtable:gtable]{gtable} (aka grob table) object.
}
\description{
Pass \code{plot_color()} R colors, and it will draw each color in a grid of
colored rectangles/cells. The background color of each cell will be the
specified colors. By default, the color name will be printed in the
foreground of the cell (in a color that contrasts with the background).

\code{plot_color()} and \code{\link[=print_color]{print_color()}} are sister functions that both display
color cells. The primary difference is where the colors are displayed: in the
console or in the graphics device.
}
\details{
Drawing is handled via the \link[gtable:gtable-package]{gtable} and
\link[grid:grid-package]{grid} packages.

If \code{col} is a vector, it's treated as a 1-column matrix. If \code{col} is a list,
then any vector within it is treated as a 1-row matrix.

If \code{col} is a named list, the names are printed in the left margin.

The foreground text color of each cell is determined by \code{\link[=bg2fg]{bg2fg()}}.
}
\examples{
plot_color(palette()) # vertical
plot_color(list(palette())) # horizontal

plot_color(matrix(rainbow(80), ncol = 8))

# Plot all the RColorBrewer palettes
if (requireNamespace("RColorBrewer")) {
  pals <- sapply(rownames(RColorBrewer::brewer.pal.info), function(pal) {
    RColorBrewer::brewer.pal(Inf, pal)
  }) |> suppressWarnings() |> rev()

  plot_color(pals, label = FALSE)
}

# Convert 'volcano' to a matrix of colors, then plot
mat <- volcano
n <- 50
color_idx <- cut(mat, breaks = n, labels = FALSE)
color_mat <- structure(rainbow(n)[color_idx], dim = dim(mat))
plot_color(color_mat, label = FALSE, index = FALSE, border = NULL)
}
\seealso{
\code{\link[=print_color]{print_color()}} for printing colors in the console.
}
