% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-color.R
\name{print_color}
\alias{print_color}
\title{Print colored cells in the console}
\usage{
print_color(col, quote = FALSE, width = NULL, gap = NULL, max = NULL)
}
\arguments{
\item{col}{A vector, matrix, or list of vectors/matrices, all containing R
colors. Can be character or numeric. The alpha (transparency) channel will
be ignored.}

\item{quote}{\code{TRUE}/\code{FALSE} for whether to print quotation marks around color
strings.}

\item{width}{Determines the width of each color cell and also whether the
color string text is printed in the blcellock's foreground. There are 3
options:
\itemize{
\item If \code{NULL} (the default), the color string text is printed, and all color
cells will be the same size. The size will be set based on the width of
the longest color string.
\item If \code{NA}, the color string text is printed, and each color cell will be
only as wide as needed to fit its text.
\item If a number >= 1, the color string text is \emph{not} printed, and each cell
will be exactly \code{floor(width)} characters wide.
}}

\item{gap}{Determines the horizontal gap between cells There are 2 options:
\itemize{
\item If \code{NULL} (the default): If \code{width} is a number (so cells have no
foreground text), there will be no gap. Otherwise (when there is
foreground text), the gap will be 1 space.
\item If a number >= 0, the gap will be \code{floor(gap)} spaces.
}}

\item{max}{Determines the maximum number of colors to print. Either \code{NULL}
(the default, meaning use \code{getOption("max.print")}) or a number >= 0. The
number will be truncated (\code{floor(max)}) and then will be honored exactly
when printing a vector or approximately when printing a matrix.}
}
\value{
Invisibly returns \code{col}.
}
\description{
Pass \code{print_color()} R colors, and it will print each color as a colored
rectangle/cell. The background color of the cell will be the specified color
(or a close approximation, depending on how many colors your console
supports). By default, the color string text will be printed in the
foreground of the cell (in a color that contrasts with the background).

\code{print_color()} and \code{\link[=plot_color]{plot_color()}} are sister functions that both display
color cells. The primary difference is where the colors are displayed: in the
console or in the graphics device.
}
\details{
\code{print_color()} works best if your console supports 24-bit colors. If it
does, then your console can display all 2^24 colors (equivalently, 256^3 or
~16.7 million colors) that can be specified using the \code{"#rrggbb"} hex RGB
syntax. If your console does not support that many colors, then it may
(depending on your platform) map each specified color to the closest
printable color.

To visually check whether your console supports 24-bit colors, run the
following code. If the result is a fairly smooth gradient within each row,
then you have 24-bit colors. If you instead see color banding (wide, distinct
steps between colors within a row) or worse (e.g., monochrome), then you have
fewer colors.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# test: can the console print in 24-bit color?
print_color(matrix(rainbow(240), nrow = 6, byrow = TRUE), width = 1)
}\if{html}{\out{</div>}}

The foreground text color of each cell is determined by \code{\link[=bg2fg]{bg2fg()}}.
}
\examples{
print_color(c("red", "steelblue", "#40E0D0"))
print_color(c("red", "steelblue", "#40E0D0"), width = 5)

print_color(rainbow(10))
print_color(rainbow(80), width = 1)

# Print all the RColorBrewer palettes
if (requireNamespace("RColorBrewer")) {
  pals <- sapply(rownames(RColorBrewer::brewer.pal.info), function(pal) {
    RColorBrewer::brewer.pal(Inf, pal)
  }) |> suppressWarnings() |> rev()

  print_color(pals, width = 3)
}

# Convert 'volcano' to a matrix of colors, then print
mat <- volcano[1:20, 1:20]
n <- 50
color_idx <- cut(mat, breaks = n, labels = FALSE)
color_mat <- structure(rainbow(n)[color_idx], dim = dim(mat))
print_color(color_mat, width = 3)
}
\seealso{
\code{\link[=plot_color]{plot_color()}} for plotting colors in the graphics device.
}
